Client/Michelson: drop support for annotations on UNPAPAIR macros
The support for annotations for the UNPAPAIR family of macros was only
partial as shown by the macro expansion test in
tests_python/contracts/macros/unpair_macro.tz (when given more than 4
annotations, the UNPPAIPAIR macros ignore the extra
annotations). Fixing this seems much work for little benefit so we
prefer to drop support for this feature.