Shell/RPCs: hide metadata from RPC responses when missing (breaking)
Modify the block related RPC services encodings to allow the absence of
block/operations metadata (i.e. in full history_mode) instead of
failing with a `Not_found`.

Warning: potential retro-compatibility break as the JSON and binary
schemas are modified. Third-party applications will need to handle the
potential absence of both the "metadata" field in the
/chains/<chain_id>/blocks/<block> and in the operation receipt in the
returned JSON object. This only affects pruned blocks so this has no
consequences on Archive and Rolling nodes.